requirements.txt
dspy-ai>=2.5.0
pandas>=2.2
python-dotenv>=1.0
